requirements.txt
# Google client library for Python
# https://developers.google.com/drive/api/quickstart/python
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# PDF library
PyPDF2
